fix: guarantee environment variables are set for shared=NO config
body: in build_pnetcdf.sh, if the shared config option is NO,
the configuration step will blow up due to an undefined
environment variable 'shared_flags'.  Likewise, if pnetcdf is
enabled for build_pio.sh, the PNETCDF env var might be undefined
if a default value is not set.  I also modified the
stack_custom.yaml file to disable the 'shared' option throughout.