feat: add container support for intel oneapi and ufs-s2s-model
body: This commit is meant to add ubuntu-based docker containers
which utilize Intel's license free compiler OneAPI along with
the latest hpc-stack.  The container support of the ufs-s2s-model
is provided as a multi-stage build such that one could re-use
the base containers for other purposes if they wish.  I have also
provided a couple readme files which help show how to setup
singularity (which can be run on-prem at least on Hera and in the
cloud) without root privilages.  Finally, some instructions on
how to convert docker images to singularity images are also
provided.