feat: add support for a unified hpc-stack and jedi-stack
body: This modification of hpc-stack enables support for compiles
of both noaa-emc/gsi and jcsda-internal/soca-science as well as
the ufs weather model.  This branch also adds a stack config which
includes compile of cmake as well as it seems like compile of
the jedi repos goes more smoothly when cmake is compiled vs when
it is untarred from a pre-built package.