Fix DOWNLOAD_ONLY option
Add a CI build that runs DOWNLOAD_ONLY. It's more difficult to test the build because the runner always has internet access, so it won't run into the `git fetch` issue.

Remove `git fetch`. I think it's a fair assumption to enforce that the downloaded packages have the required tags from the original clone.

Download package before running nc-config for packages that use it.