Change load, try_load, and prereq to depends_on
depends_on is the correct behavior in most circumstances. If the dependent module is already loaded it will use that, but if not will try to load it. This way users can mix and match their own versions without LMod changing their selection.

depends_on combines the functionality of load/preqreq in a single step.

In some cases such as MAPL, compiler/MPI meta modules, and HDF5 with szip load/try_load where the module is has to exist or might not exist load/try_load is the correct behavior.